challenge_api/requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==1.0.0
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
drf-yasg==1.21.4
environs==9.5.0
exceptiongroup==1.0.0rc9
h11==0.14.0
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.18.0
outcome==1.2.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.4
py==1.11.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.3
pytest-django==4.5.2
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.12
wsproto==1.2.0
